suprise! time for hive based auth,  default user is root, hive API perms, and telepath UNIX sockets, a cortex layer API and cortex view plumbing (#1090)
* Update hiveauth support for managing all user auth in a cell
* unix socket support for telepath
* add remote layer reconnection logic
* Update cortex view logic to allow multiple layers to be added to a view
* added compat test for cell auth migration
* robustify the cortex server tool test
* Bump the pre-release version number.
* remove misc dead code and imports

+++ synapse/tools/backup.py
@@ -41,6 +41,10 @@ def backup(srcdir, dstdir):
 
         for name in fnames:
             srcpath = s_common.genpath(root, name)
+            # skip unix sockets etc...
+            if not os.path.isfile(srcpath):
+                continue
+
             dstpath = s_common.genpath(dstdir, relpath, name)
             logger.info(f'copying: {srcpath} -> {dstpath}')
             shutil.copy(srcpath, dstpath)
